composerv2 compatibility
