+++ composer.json
@@ -14,10 +14,10 @@
     "require": {
         "php": ">=7.1",
         "ext-json": "*",
-        "composer-plugin-api": "^1.1.0"
+        "composer-plugin-api": "^2"
     },
     "require-dev": {
-        "composer/composer": "^1.8.4",
+        "composer/composer": "^2",
         "dealerdirect/phpcodesniffer-composer-installer": "^0.5",
         "object-calisthenics/phpcs-calisthenics-rules": "^3.5",
         "slevomat/coding-standard": "^5.0",
